refactor: move module overlap error rendering function to `module-overlap.ts`
